refactor(auth): fold blob-export entitlement check into authorizeBlobStorageRequest
